align couchbase example to other samples
