+++ couchbase-persistence/couchbase-persistence-java-mvn/hello-api/src/main/java/com/lightbend/lagom/samples/hello/api/GreetingMessage.java
@@ -1,0 +1,20 @@
+package com.lightbend.lagom.sampleshello.api;
+
+import lombok.Value;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.google.common.base.MoreObjects;
+import com.google.common.base.Preconditions;
+
+@Value
+@JsonDeserialize
+public final class GreetingMessage {
+
+  String message;
+
+  @JsonCreator
+  public GreetingMessage(String message) {
+    this.message = Preconditions.checkNotNull(message, "message field MUST not be null");
+  }
+}
