--- couchbase-persistence/couchbase-persistence-java-mvn/hello-api/src/main/java/org/example/hello/api/UserGreeting.java
@@ -1,16 +1,0 @@
-package org.example.hello.api;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import lombok.Value;
-
-@Value
-public final class UserGreeting {
-  String user;
-  String message;
-
-  @JsonCreator
-  public UserGreeting(String user, String message) {
-    this.user = user;
-    this.message = message;
-  }
-}
