--- couchbase-persistence/couchbase-persistence-java-mvn/hello-impl/src/main/java/org/example/hello/impl/HelloModule.java
@@ -1,13 +1,0 @@
-package org.example.hello.impl;
-
-import com.google.inject.AbstractModule;
-import com.lightbend.lagom.javadsl.server.ServiceGuiceSupport;
-import org.example.hello.api.HelloService;
-
-public class HelloModule extends AbstractModule implements ServiceGuiceSupport {
-
-  @Override
-  protected void configure() {
-    bindService(HelloService.class, HelloServiceImpl.class);
-  }
-}
