+++ couchbase-persistence/couchbase-persistence-java-mvn/hello-impl/src/main/resources/application.conf
@@ -1,4 +1,4 @@
-play.modules.enabled += org.example.hello.impl.HelloModule
+play.modules.enabled += com.lightbend.lagom.sampleshello.impl.HelloModule
 
 lagom.persistence.ask-timeout = 10s
 akka.cluster.sharding.state-store-mode = ddata
