+++ couchbase-persistence/couchbase-persistence-scala-sbt/hello-impl/src/main/resources/application.conf
@@ -1,4 +1,4 @@
-play.application.loader = org.example.hello.impl.HelloLoader
+play.application.loader = com.lightbend.lagom.sampleshello.impl.HelloLoader
 play.allowGlobalApplication = true
 
 akka.cluster.sharding.state-store-mode = ddata
